pyproject.toml
[project]
name = "email-resend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "python-dotenv>=1.0.1",
    "resend>=2.6.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.394",
    "ruff>=0.9.6",
]
